Update spritesheet; partially implement laser eye
Drew a new set of cave tiles (which fit together much more easily than
the old ones) and some various edges.

Spritesheet is a bit of a mess, still, since the original tile is still
in and there was no room around it to fit the new tiles; I'm dooming
myself to a huge tile rearrangement later.

Laser eye is mostly non-functional.  Need to spruce up some physics
stuff a bit before I finish it.