Vastly improve handling of rounding in collision, I hope
I basically deleted all the rounding code and everything is much better.
The gradual sliding uphill and a couple of clipping problems are all
gone.  Go figure.

The actor code still tries to move to the nearest pixel, so if you're
standing on flat ground, you'll probably be pixel-aligned.  If you're
standing on a slope, you'll have some awful float coordinates, but
apparently that's fine.